Finished Layer 1 Work
I added a transition manager prefab to go in every scene. With this you can call its function LoadScene to load to a differnet level. All you have to do is pass in the integer of the scene index you want to go to. There is also a test scenario set up where each to you go to the scene test(green lady) you will be teleported to a differnt scene until you cycle through all of them, which it will then no longer switch your scene. Also added a basic movement system for the player with a basic sprite.